requirements.txt
aiogram==3.4.1
openai-whisper==20231117
python-docx==1.1.0
aiohttp==3.9.5
ffmpeg-python==0.2.0
requests==2.31.0
numpy>=1.24.0
pydub==0.25.1
tqdm==4.66.4
asyncpg
python-dotenv
